% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{CreativeStatConnection}
\alias{CreativeStatConnection}
\title{Connecting to the database MySql (Creative database stats)}
\usage{
CreativeStatConnection()
}
\value{
None
}
\description{
Creates connection to the database (Creative database stats)). To connect the function uses Renviron 
(read more \code{\link{ReadRemoteEnvironment}})
}
\examples{
creative <- CreativeStatConnection()
query <- 'SELECT * FROM table LIMIT 1'
campaigns <- dbGetQuery(creative, query)
dbDisconnect(creative)

}
\seealso{
\code{\link{ReadRemoteEnvironment}}
}
