% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{GaleraReadNoLimitConnection}
\alias{GaleraReadNoLimitConnection}
\title{Connecting to the database MySql (GaleraReaOnly with an increased limit)}
\usage{
GaleraReadNoLimitConnection()
}
\value{
None
}
\description{
Creates connection to the database (GaleraReaOnly). To connect the function uses Renviron 
(read more \code{\link{ReadRemoteEnvironment}})
}
\examples{
galera <- GaleraReadNoLimitConnection()
query <- 'SELECT * FROM table LIMIT 1'
campaigns <- dbGetQuery(galera, query)
dbDisconnect(galera)

}
\seealso{
\code{\link{ReadRemoteEnvironment}}
}
