% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{GaleraWriteConfigConnection}
\alias{GaleraWriteConfigConnection}
\title{Connecting to the database MySql (GaleraWrite - configuration regression)}
\usage{
GaleraWriteConfigConnection()
}
\value{
None
}
\description{
Creates connection to the database (GaleraWrite - configuration regression). To connect the function uses Renviron 
(read more \code{\link{ReadRemoteEnvironment}})
}
\examples{
galera <- GaleraWriteConfigConnection()
query <- 'SELECT * FROM table LIMIT 1'
campaigns <- dbGetQuery(galera, query)
dbDisconnect(galera)

}
\seealso{
\code{\link{ReadRemoteEnvironment}}
}
