% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{GaleraWriteExperementConnection}
\alias{GaleraWriteExperementConnection}
\title{Connecting to the database MySql (GaleraWrite - experement regression)}
\usage{
GaleraWriteExperementConnection()
}
\value{
None
}
\description{
Creates connection to the database (GaleraWrite - experement regression). To connect the function uses Renviron 
(read more \code{\link{ReadRemoteEnvironment}})
}
\examples{
galera <- GaleraWriteExperementConnection()
query <- 'SELECT * FROM table LIMIT 1'
campaigns <- dbGetQuery(galera, query)
dbDisconnect(galera)

}
\seealso{
\code{\link{ReadRemoteEnvironment}}
}
