% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-gitlogo.R
\docType{data}
\name{git_logo}
\alias{git_logo}
\alias{github}
\alias{gitlab}
\alias{bitbucket}
\title{Example Images}
\format{
An object of class \code{magick-image} of length 1.
}
\usage{
github
}
\description{
Example images included with `bunny`:
}
\details{
- `github`: Github Logo, b/w, 512x500
 - `gitlab`: Gitlab Logo, b/w, 512x474
 - `bitbucket` : Bitbucket Logo, b/w, 512x512
}
\keyword{datasets}
