% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-channel.R
\name{image_alpha_drop}
\alias{image_alpha_drop}
\title{Drop matte layer}
\usage{
image_alpha_drop(image)
}
\arguments{
\item{image}{image with opacity}
}
\value{
image without opacity layer
}
\description{
Easily remove alpha/opacity/matter layer from image
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
