% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-canvas.R
\name{image_canvas_ghcard}
\alias{image_canvas_ghcard}
\alias{image_border_ghcard}
\title{Make your own github card}
\usage{
image_canvas_ghcard(fill_color = "white", width = 1280, height = 640,
  border = "80x80")

image_border_ghcard(image, border_color = "white", border = "80x80")
}
\arguments{
\item{fill_color}{background color of the card. Default: 'white'}

\item{width}{card width (including border). Default: 1280}

\item{height}{card height (including border). Default: 640}

\item{border}{border size (will be deducted from the canvas). Default: '80x80'}

\item{image}{completed gihub card to be complemented by border}

\item{border_color}{color of the border. Preferrably same as `fill_color`. Default: 'white'}
}
\value{
github card of specified dimensions
}
\description{
Produces canvas for assembling github social media card of standard size
}
\details{
The card excludes borders to make sure you never have anything trimmed. Border can be added with `image_border_ghcard()`, which is just a thin wrapper over `magick::image_border()`
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
