% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-canvas.R
\name{image_canvas_hex}
\alias{image_canvas_hex}
\alias{image_canvas_hexborder}
\title{Make your own hex sticker}
\usage{
image_canvas_hex(border_color = "black", border_size = 2,
  fill_color = "azure", outer_margin = "10x10", scope = FALSE,
  scope_color = "grey80")

image_canvas_hexborder(border_color = "black", border_size = 2,
  outer_margin = "10x10")
}
\arguments{
\item{border_color}{color of the hex border. Default: 'black'}

\item{border_size}{border thickness. Default: 2}

\item{fill_color}{hex background. Default: 'azure'}

\item{outer_margin}{margin around the hex. May be helpful for making sure nothing gets trimmed. Default: '10x10'}

\item{scope}{logical. Add measurement grid to the center of the hex (similar to rifle scope).}

\item{scope_color}{color of the grid. Default "grey80".}
}
\value{
image of hex canvas
}
\description{
Produces canvas and border for assembling hex sticker
}
\details{
Border around the sticker can be later trimmed either by the printing company or by `magick::image_trim()`. It is preferred to not choose white background or white border to prevent excessive trimming.
Requires `ggplot` and `ggforce` installed. `image_canvas_hexborder()` produces transparent hex with specified border color.
}
\examples{
\dontrun{
if(interactive()){
 image_canvas_hex()
 }
}
}
\seealso{
\code{\link[magick]{transform}},\code{\link[magick]{color}},\code{\link[magick]{attributes}},\code{\link[magick]{painting}}
}
