% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-channel.R
\name{image_channel_reset}
\alias{image_channel_reset}
\alias{image_channel_copy}
\title{Copy or reset a selected channel}
\usage{
image_channel_reset(image, channel)

image_channel_copy(image, channel, channel_image)
}
\arguments{
\item{image}{image}

\item{channel}{channel name}

\item{channel_image}{image containing channel to copy from}
}
\value{
image with modified channel
}
\description{
Copy a content of a selected channel from another image or reset a selected channel to black
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
