% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-channel.R
\name{image_channel_combine}
\alias{image_channel_combine}
\alias{image_channel_separate}
\title{Combine or separate channel images}
\usage{
image_channel_combine(image, channel_index = "RGB", canvas = NULL)

image_channel_separate(image)
}
\arguments{
\item{image}{image stack (for `image_channel_combine()`) or individual image (for `image_channel_separate()`)}

\item{channel_index}{character vector with channel names (or abbreviation). Default: 'RGB'. Images in the stack will be interpreted as set of channels in the specified sequence.}

\item{canvas}{image to be used as canvas to start copying channels into, Default: NULL}
}
\value{
single (flat) image with channels copied from the image stack
}
\description{
Split image by channel, creating a stack of images or combine a stack of grayscale images into a full color image
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
