% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-compose.R
\name{image_compose}
\alias{image_compose}
\title{Combine two images together using operators, gravity and offset}
\usage{
image_compose(
  image,
  composite_image,
  operator = "atop",
  gravity = "NorthWest",
  offset = "+0+0",
  compose_args = ""
)
}
\arguments{
\item{image}{original image which will remain in the background (also known as "destination" image)}

\item{composite_image}{image place on top of it (aslo known as "source" image)}

\item{operator}{magick operator. See more at [ImageMagick compose](https://www.imagemagick.org/Usage/compose/) page. Default: 'atop'}

\item{gravity}{one of `magick::gravity_types()`.  Default: 'NorthWest'}

\item{offset}{geometry point specification of offset. Affected by choice of `gravity`. Default: '+0+0'}

\item{compose_args}{other parameters required by operator. Default: ''}
}
\value{
combined image
}
\description{
Extends `image_composite()` to include `gravity` argument which allows targeted placement of images
}
\examples{
\dontrun{
if(interactive()){
 frink <- image_read("https://jeroen.github.io/images/frink.png")
 image_compose(frink, magick::rose, gravity="North")
 }
}
}
\seealso{
\code{\link[magick]{attributes}},\code{\link[magick]{composite}}
 \code{\link[bunny]{geometry_parse}}
}
