% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-crop.R
\name{image_shave}
\alias{image_shave}
\title{Remove border from image}
\usage{
image_shave(image, geometry)
}
\arguments{
\item{image}{original image}

\item{geometry}{geometry specified by area format, for example "10x10".}
}
\value{
"shaved" image
}
\description{
Reverse operation to `magick::image_border()` removing area aroud the image specified by area geometry
}
\examples{
\dontrun{
if(interactive()){
 magick::rose \%>\%
 image_border() \%>\%
 image_shave("10x10")
 }
}
}
\seealso{
\code{\link[magick]{attributes}},\code{\link[magick]{transform}}
}
