% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-pixels.R
\name{image_getpixel}
\alias{image_getpixel}
\title{Extract pixel color using specified point geometry}
\usage{
image_getpixel(img, geometry)
}
\arguments{
\item{img}{magick image}

\item{geometry}{point geomentry (generated by `magick::geometry_point()`) in the format (`+x+y`)}
}
\value{
hex value of the color
}
\description{
Extracts hex color of the pixel at the specified location
}
\examples{
\dontrun{
if(interactive()){
 frink <- image_read("https://jeroen.github.io/images/frink.png")
 image_getpixel(frink, geometry_point(100,100))
 }
}
}
\seealso{
\code{\link[magick]{editing}},\code{\link[magick]{transform}}
}
