% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-canvas.R, R/oxy-info.R
\name{image_canvas}
\alias{image_canvas}
\alias{image_annotate_info}
\title{Empty canvas from existing image or specified dimensions}
\usage{
image_canvas(
  image = NULL,
  color = "transparent",
  width = NULL,
  height = NULL,
  units = "px",
  dpi = 96,
  pseudo_image = ""
)

image_annotate_info(
  text,
  size = 10,
  font = "",
  style = "normal",
  weight = 400,
  ...
)
}
\arguments{
\item{image}{existing image to copy dimensions from (content will be discarded)}

\item{color}{color of the blank image. Default: 'transparent'}

\item{width}{canvas width}

\item{height}{canvas width}

\item{units}{one of "in", "cm" or "px". Default="px".}

\item{dpi}{specification of density (required for any `units` other than "px"). Default = 96}

\item{pseudo_image}{name of the pseudo-image. Default: ""}

\item{text}{proposed text for annotation}

\item{size}{text size. Default: 10}

\item{font}{text font to be passed to `image_annotate()`. Default: ''}

\item{style}{font style to be passed to `image_annotate()` (e.g. "normal", "italic", "oblique").  Default: 'normal'}

\item{weight}{font weight. Normal is 400, bold is 700. Default: 400}

\item{...}{currently ignored}
}
\value{
blank image of given dimensions

data frame with measurements for every letter and the phrase in total
}
\description{
Extends `magick::image_blank()` to allow creating blank image with or without manually specifying width and height. If specimen image is provided dimensions will be copied from it.

Calculate space taken by every latter in annotation
}
\details{
The purpose of this function is to calculate kerning between letters that fonts introduce on the fly. This can be helpful in designing custom kerning.
}
\examples{
\dontrun{
if(interactive()){
 frink <- image_read("https://jeroen.github.io/images/frink.png")
 image_canvas(frink, "pink")

 # will produce 504x360 canvas
 image_canvas(width=7, height=5, units="in", dpi=72, color="burlywood")
 # will produce 800x600 canvas
 image_canvas(width=800, height=600, color="cornsilk")

 # will produce 377x566 canvas
 image_canvas(width=10, height=15, units="cm", dpi=96, color="aliceblue")
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
