% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-draw.R
\name{image_plot}
\alias{image_plot}
\title{Draw a point or a rectangle on image}
\usage{
image_plot(img, geometry, col = "green", pointsize = 5, ...)
}
\arguments{
\item{img}{magick image}

\item{geometry}{point or area geometry}

\item{col}{character value of color (either name or hex value), Default: 'green'}

\item{pointsize}{size of point, Default: 5}

\item{...}{other parameters passed to `image_draw()`}
}
\value{
magick image with drawing on it
}
\description{
Draws pixel or rectangle on top of specified {magick} image
}
\details{
Function makes distinction between drawing a point or drawing a rectangle only by the geometry. Use `geometry_point()` and `geometry_area()` to prepare and format geometries.
}
\examples{
\dontrun{
if(interactive()){
 frink <- image_read("https://jeroen.github.io/images/frink.png")
 image_plot(frink, geometry_area(50,35,80,150))
 image_plot(frink, geometry_point(70,70))
 }
}
}
\seealso{
\code{\link[magick]{device}}
}
