% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{InfluxAuthenticateParams}
\alias{InfluxAuthenticateParams}
\title{Authentication in influx}
\usage{
InfluxAuthenticateParams()
}
\value{
None
}
\description{
To connect the function uses Renviron 
(read more \code{\link{ReadRemoteEnvironment}})
}
\examples{
influx <- InfluxAuthenticateParams()

}
\seealso{
\code{\link{ReadRemoteEnvironment}}
}
