% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{kMonitorConnectionParams}
\alias{kMonitorConnectionParams}
\title{Authentication in monitor}
\usage{
kMonitorConnectionParams()
}
\value{
None
}
\description{
To connect the function uses Renviron 
(read more \code{\link{ReadRemoteEnvironment}})
}
\examples{
monitor <- kMonitorConnectionParams()

}
\seealso{
\code{\link{ReadRemoteEnvironment}}
}
