% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{PostrgresAuditReadConnection}
\alias{PostrgresAuditReadConnection}
\title{Connecting to the database Postgres (Adp Audit)}
\usage{
PostrgresAuditReadConnection()
}
\value{
None
}
\description{
Creates connection to the database (Adp Audit). To connect the function uses Renviron 
(read more \code{\link{ReadRemoteEnvironment}})
}
\examples{
adp <- PostrgresAuditReadConnection()
query <- 'SELECT * FROM table LIMIT 1'
campaigns <- dbGetQuery(adp, query)
dbDisconnect(adp)

}
\seealso{
\code{\link{ReadRemoteEnvironment}}
}
