% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{ReadRemoteEnvironment}
\alias{ReadRemoteEnvironment}
\title{Read Renviron
This function downloads the Renviron file to the home directory.}
\usage{
ReadRemoteEnvironment(user)
}
\arguments{
\item{user}{user name on server r2d2}
}
\value{
None (~/ downliad file '.Renviron')
}
\description{
Read Renviron
This function downloads the Renviron file to the home directory.
}
\examples{
ReadRemoteEnvironment('i.ivanov')

}
