% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-hough.R
\name{tidy_hough_mvg}
\alias{tidy_hough_mvg}
\title{Tidy up Hough Lines object}
\usage{
tidy_hough_mvg(mvg, margin = 0.05)
}
\arguments{
\item{mvg}{Magick Vector Graphics object, as output by `magick::image_hough_txt(format='mvg')``}

\item{margin}{Allowed margin around the image for detection of line intersections. Default: 0.05}
}
\value{
A list containing Hough Lines parameters used, number of lines and intersections detected, detailed data about lines and intersections.
}
\description{
Create a list summarizing hough lines and their intersections
}
\examples{
\dontrun{
if(interactive()){
 if(magick_config()$version > "6.8.9"){
   shape <- demo_image("shape_rectangle.gif")
   rectangle <- image_canny(shape)
   rectangle \%>\% image_hough_txt(format = 'mvg') \%>\% tidy_hough_mvg()
   }
 }
}
}
