% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{VerticaFullStatsConnection}
\alias{VerticaFullStatsConnection}
\title{Connecting to the database Vertic (Full stats)}
\usage{
VerticaFullStatsConnection()
}
\value{
None
}
\description{
Creates connection to the database (full stats)
}
\examples{
vertica <- VerticaRnaConnection()
query <- 'SELECT * FROM public.ox_campaigns LIMIT 1'
campaigns <- dbGetQuery(vertica, query)
dbDisconnect(vertica)

}
\seealso{
\code{\link{ReadRemoteEnvironment}}
}
