% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{VerticaRealStatsConnection}
\alias{VerticaRealStatsConnection}
\title{Connecting to the database Vertica (Realtime stats)}
\usage{
VerticaRealStatsConnection()
}
\value{
None
}
\description{
Creates connection to the database (Realtime stats). To connect the function uses Renviron 
(read more \code{\link{ReadRemoteEnvironment}})
}
\examples{
vertica <- VerticaRealStatsConnection()
query <- 'SELECT * FROM public.table LIMIT 1'
campaigns <- dbGetQuery(vertica, query)
dbDisconnect(vertica)

}
\seealso{
\code{\link{ReadRemoteEnvironment}}
}
