% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{VerticaRnaConnection}
\alias{VerticaRnaConnection}
\title{Connecting to the database Vertica (ETL-RNA)}
\usage{
VerticaRnaConnection()
}
\value{
None
}
\description{
Creates connection to the database (ETL-RNA). To connect the function uses Renviron 
(read more \code{\link{ReadRemoteEnvironment}})
}
\examples{
vertica <- VerticaRnaConnection()
query <- 'SELECT * FROM public.table LIMIT 1'
campaigns <- dbGetQuery(vertica, query)
dbDisconnect(vertica)

}
\seealso{
\code{\link{ReadRemoteEnvironment}}
}
