% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-canvas.R, R/oxy-info.R
\name{image_canvas}
\alias{image_canvas}
\alias{image_annotate_info}
\title{Empty canvas from existing image}
\usage{
image_canvas(img, color = "transparent", pseudo_image = "")

image_annotate_info(text, size = 10, font = "", style = "normal",
  weight = 400, ...)
}
\arguments{
\item{img}{existing image to copy dimensions from (content will be discarded)}

\item{color}{color of the blank image. Default: 'transparent'}

\item{pseudo_image}{name of the pseudo-image. Default: ''}

\item{text}{proposed text for annotation}

\item{size}{text size. Default: 10}

\item{font}{text font to be passed to `image_annotate()`. Default: ''}

\item{style}{font style to be passed to `image_annotate()` (e.g. "normal", "italic", "oblique").  Default: 'normal'}

\item{weight}{font weight. Normal is 400, bold is 700. Default: 400}

\item{...}{currently ignored}
}
\value{
blank image of given dimensions

data frame with measurements for every letter and the phrase in total
}
\description{
Allows creating blank image without manually specifying width and height. These dimensions will be copied from the image passed as first argument.

Calculate space taken by every latter in annotation
}
\details{
The purpose of this function is to calculate kerning between letters that fonts introduce on the fly. This can be helpful in designing custom kerning.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
