% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{VerticaFullStatsConnection}
\alias{VerticaFullStatsConnection}
\title{Connecting to the database Vertica (Full stats)}
\usage{
VerticaFullStatsConnection()
}
\value{
None
}
\description{
Creates connection to the database (full stats). To connect the function uses Renviron 
(read more \code{\link{ReadRemoteEnvironment}})
}
\examples{
vertica <- VerticaFullStatsConnection()
query <- 'SELECT * FROM public.table LIMIT 1'
campaigns <- dbGetQuery(vertica, query)
dbDisconnect(vertica)

}
\seealso{
\code{\link{ReadRemoteEnvironment}}
}
